5_AISearch/vector_search/requirements.txt
azure-search-documents
azure-identity
openai